--- toio_mcp/__init__.py
@@ -1,5 +1,0 @@
-"""
-toio-mcp - MCP server for toio Core Cube
-"""
-
-__version__ = "0.1.0"
